LibI4/Python_AI/Python_Plugins/Extra/requirements.txt
voicevox-client
speechrecognition
asyncio
pygame
discord.py
websockets
requests